# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PeAPIniere API",
        "description": "API for managing plant statistics, categories, and orders in PeAPIniere",
        "contact": {
            "name": "Support Team",
            "email": "support@peapiniere.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com"
        }
    ],
    "paths": {
...
        "/api/plants": {
            "get": {
                "tags": [
                    "Plant"
                ],
                "summary": "Display a listing of the plants",
                "operationId": "6e2a3a91967b1e364c87b755cd5152f6",
                "responses": {
                    "200": {
                        "description": "List of plants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "category_id": {
                                                "type": "integer"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "nullable": true
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "slug": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Plant"
                ],
                "summary": "Store a newly created plant in storage",
                "operationId": "d005e00370ab9bd321d50ba5dde92042",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plant created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "plant": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "category_id": {
                                                    "type": "integer"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "nullable": true
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/plants/{slug}": {
            "get": {
                "tags": [
                    "Plant"
                ],
                "summary": "Display the specified plant",
                "operationId": "cf101271037916b9c4a3e8bed0a3bf4e",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of the plant",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The specified plant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "category_id": {
                                            "type": "integer"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "nullable": true
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "slug": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plant not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Plant"
                ],
                "summary": "Update the specified plant",
                "operationId": "e5981af9ce55cca7fc8eee04c3c0b488",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of the plant",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plant updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "plant": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "category_id": {
                                                    "type": "integer"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "nullable": true
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plant not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Plant"
                ],
                "summary": "Remove the specified plant from storage",
                "operationId": "72abaa41b86e9d3386648d6c79c2b1ab",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of the plant",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plant deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plant not found"
                    }
                }
            }
        },
...
    },
...
}